requirements-minimal.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
psycopg2-binary==2.9.7
python-socketio==5.8.0
eventlet==0.33.3
gunicorn==21.2.0
PyJWT==2.8.0
